feat: move ObjFile components to Mesh
